backend/requirements.txt
# Web Framework
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-CORS==4.0.0

# Database
PyMySQL==1.1.0
SQLAlchemy==2.0.19

# OpenStack Clients
python-keystoneclient==5.1.0
python-novaclient==18.3.0
python-cinderclient==9.3.0
python-neutronclient==11.2.0
python-glanceclient==4.5.0
openstacksdk==1.3.1

# Security and Crypto
PyJWT==2.8.0
cryptography==41.0.7
Werkzeug==2.3.7

# SSH and Network
paramiko==3.3.1

# Data Processing  
numpy<2.0.0
pandas==2.1.1
openpyxl==3.1.2

# Utilities
PyYAML==6.0.1
python-dotenv==1.0.0
redis==4.6.0

# Timezone
pytz==2023.3

# Development Tools
pytest==7.4.2
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0